+++ Numerics/Geometry/Vectors/Vector2i.cs
@@ -847,7 +847,7 @@ namespace Ibasa.Numerics.Geometry
 		/// </summary>
 		/// <param name="value">A vector.</param>
 		/// <returns>The absolute value of value.</returns>
-		public static float Absolute(Vector2i value)
+		public static double Absolute(Vector2i value)
 		{
 			return Functions.Sqrt(AbsoluteSquared(value));
 		}
@@ -856,16 +856,16 @@ namespace Ibasa.Numerics.Geometry
 		/// </summary>
 		/// <param name="value">A vector.</param>
 		/// <returns>The normalized value of value.</returns>
-		public static Vector2f Normalize(Vector2i value)
+		public static Vector2d Normalize(Vector2i value)
 		{
 			var absolute = Absolute(value);
-			if(absolute <= float.Epsilon)
+			if(absolute <= double.Epsilon)
 			{
 				return Vector2i.Zero;
 			}
 			else
 			{
-				return (Vector2f)value / absolute;
+				return (Vector2d)value / absolute;
 			}
 		}
 		#endregion
