+++ Numerics/Geometry/Vectors/Vector2ul.cs
@@ -827,7 +827,7 @@ namespace Ibasa.Numerics.Geometry
 		/// <param name="value">A vector.</param>
 		/// <returns>The absolute value of value.</returns>
 		[CLSCompliant(false)]
-		public static float Absolute(Vector2ul value)
+		public static double Absolute(Vector2ul value)
 		{
 			return Functions.Sqrt(AbsoluteSquared(value));
 		}
@@ -837,16 +837,16 @@ namespace Ibasa.Numerics.Geometry
 		/// <param name="value">A vector.</param>
 		/// <returns>The normalized value of value.</returns>
 		[CLSCompliant(false)]
-		public static Vector2f Normalize(Vector2ul value)
+		public static Vector2d Normalize(Vector2ul value)
 		{
 			var absolute = Absolute(value);
-			if(absolute <= float.Epsilon)
+			if(absolute <= double.Epsilon)
 			{
 				return Vector2ul.Zero;
 			}
 			else
 			{
-				return (Vector2f)value / absolute;
+				return (Vector2d)value / absolute;
 			}
 		}
 		#endregion
