+++ Numerics/Geometry/Vectors/Vector4ui.cs
@@ -3994,7 +3994,7 @@ namespace Ibasa.Numerics.Geometry
 		/// <param name="value">A vector.</param>
 		/// <returns>The absolute value of value.</returns>
 		[CLSCompliant(false)]
-		public static float Absolute(Vector4ui value)
+		public static double Absolute(Vector4ui value)
 		{
 			return Functions.Sqrt(AbsoluteSquared(value));
 		}
@@ -4004,16 +4004,16 @@ namespace Ibasa.Numerics.Geometry
 		/// <param name="value">A vector.</param>
 		/// <returns>The normalized value of value.</returns>
 		[CLSCompliant(false)]
-		public static Vector4f Normalize(Vector4ui value)
+		public static Vector4d Normalize(Vector4ui value)
 		{
 			var absolute = Absolute(value);
-			if(absolute <= float.Epsilon)
+			if(absolute <= double.Epsilon)
 			{
 				return Vector4ui.Zero;
 			}
 			else
 			{
-				return (Vector4f)value / absolute;
+				return (Vector4d)value / absolute;
 			}
 		}
 		#endregion
